alog.msg -> log_msg Compiler directives Logger timer (time from program start) Res.genUid() -> objectIdGenerator, bufferIdGenerator No warnings
git-svn-id: svn://localhost/steel.new/trunk@152 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/math/line.cpp
@@ -20,7 +20,7 @@ bool isCross(const Line a, const Line b, float &t, float &s)
 	double len2 = bbxab.getSquaredLengthd();
 	if(len2<EPSILON2) return false; // линии параллельны
 	
-	s = (((a.base - b.base)*a.a)&bbxab / len2);
+	s = (((a.base - b.base)*a.a)&bbxab / (float)len2);
 
 	v3 p = b.point(s);
 
@@ -40,7 +40,7 @@ bool isCrossFast(const Line a, const Line b, float &s)
 	v3 bbxab = (b.a*a.a);
 	double len2 = bbxab.getSquaredLengthd();
 	
-	s = ((a.base - b.base)*a.a)&bbxab / len2;
+	s = ((a.base - b.base)*a.a)&bbxab / (float)len2;
 
 	return true;
 }
